+++ resources/index.html.php
@@ -32,6 +32,9 @@
             }
             .bash > span { font-family: monospace, monospace; }
             mark.default { background: rgba(220,220,220,0.75); color: #212121; }
+            .img-valign {
+              vertical-align: middle;
+            }
         </style>
     </head>
     <body>
@@ -39,7 +42,14 @@
             <div class="row">
                 <div class="col-sm-12 col-md-12 col-lg-10 col-lg-offset-1">
                     <div class="title">
-                        <h1>Packagist Mirror <img src="https://cdnjs.cloudflare.com/ajax/libs/flag-icon-css/3.1.0/flags/4x3/<?= $countryCode; ?>.svg" title="<?= $countryName; ?>" alt="<?= $countryName; ?>"/></h1>
+                        <h1>
+                            Packagist Mirror
+                            <img src="https://cdnjs.cloudflare.com/ajax/libs/flag-icon-css/3.1.0/flags/4x3/<?= $countryCode; ?>.svg"
+                                    title="<?= $countryName; ?>"
+                                    alt="<?= $countryName; ?>"
+                                    class="img-valign"
+                                    />
+                        </h1>
                         <p><span id="lastsynced" ></span><br>(Synchronized every <?= $synced ?> seconds)</p>
                     </div>
                     <p>